backend/requirements.txt
# =============================================================================
# Loyallia Backend — Python Dependencies (all versions pinned)
# =============================================================================

# Core Django
django==5.0.6
django-ninja==1.3.0
email-validator==2.2.0
django-cors-headers==4.4.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Cache & Queue
redis==5.0.7
django-redis==5.4.0
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
kombu==5.3.7

# Async task monitoring
flower==2.0.1

# File storage (MinIO / S3-compatible)
django-storages==1.14.4
boto3==1.34.144
botocore==1.34.144

# Image processing & QR
Pillow==10.4.0
qrcode[pil]==7.4.2

# Cryptography & JWT
PyJWT==2.8.0
cryptography==43.0.0
argon2-cffi==23.1.0
passlib==1.7.4

# Apple Wallet PKPass generation
# Using standard library + cryptography for signing
pyopenssl==24.2.1

# Push Notifications
firebase-admin==6.5.0
google-auth==2.29.0          # FCM HTTP v1 API token exchange (also used by firebase-admin)
google-auth-httplib2==0.2.0  # Required transport for google-auth
httpx[http2]==0.27.2         # APNs HTTP/2 + general HTTP client

# Claro Pay Ecuador (uses httpx — already listed above)

# Email
django-anymail==11.1

# Configuration
python-decouple==3.8

# WSGI server
gunicorn==22.0.0

# Utilities
python-slugify==8.0.4
shortuuid==1.0.13
Markdown==3.6

# Data Processing
pandas==2.2.2
openpyxl==3.1.2
xlrd==2.0.1

# PDF generation (invoices)
reportlab==4.2.2

# Development only (not in production image, but keep for consistency)
django-extensions==3.2.3
